requirements.txt
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cloudpickle==0.6.1
cycler==0.10.0
dask==1.0.0
distributed==1.25.1
future==0.17.1
HeapDict==1.0.0
idna==2.8
kiwisolver==1.0.1
matplotlib==3.0.2
msgpack==0.6.0
numpy==1.15.4
pandas==0.23.4
patsy==0.5.1
psutil==5.4.8
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
sortedcontainers==2.1.0
statsmodels==0.9.0
tblib==1.3.2
toolz==0.9.0
tornado==5.1.1
tqdm==4.28.1
tsfresh==0.11.1
urllib3==1.24.1
zict==0.1.3
